GPS fixes

+++ LVBox/GPS.h
@@ -2,12 +2,13 @@
 #define GPS_H
 #include <stdint.h>
 
-extern double gps_longitude;
-extern double gps_latitude;
+extern float gps_longitude;
+extern float gps_latitude;
 extern float gps_speed;
 
+/// Initialize the GPS
 void gps_init();
-
+/// Update the GPS to get new data
 void gps_poll();
 
 #endif
